# [D] Apply flat-list approach