Can we also check flat-list approach for D?