@jinyus 2023-11-18
Much slower...same for C#, Java and Rust Con.

Main: `25.45 ms  | 325.21 ms  | 2.79 s `

This PR:
```
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 40.581ms
        Processing time (w/o IO): 41.002ms
        Processing time (w/o IO): 40.934ms
        Processing time (w/o IO): 40.94ms
        Processing time (w/o IO): 40.801ms
        Processing time (w/o IO): 40.756ms
        Processing time (w/o IO): 40.868ms
        Processing time (w/o IO): 40.875ms
        Processing time (w/o IO): 40.916ms
        Processing time (w/o IO): 40.884ms
        Processing time (w/o IO): 40.837ms
        Processing time (w/o IO): 40.778ms
        Processing time (w/o IO): 40.863ms
          Time (mean ± σ):     121.7 ms ±   1.0 ms    [User: 105.9 ms, System: 17.1 ms]
          Range (min … max):   120.5 ms … 123.6 ms    10 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 566.261ms
        Processing time (w/o IO): 564.796ms
        Processing time (w/o IO): 565.202ms
          Time (mean ± σ):     873.1 ms ±   0.6 ms    [User: 843.6 ms, System: 33.6 ms]
          Range (min … max):   872.8 ms … 873.5 ms    2 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 4950.49ms
        Processing time (w/o IO): 4953.9ms
        Processing time (w/o IO): 4952.95ms
          Time (mean ± σ):      5.989 s ±  0.006 s    [User: 5.846 s, System: 0.151 s]
          Range (min … max):    5.985 s …  5.993 s    2 runs

```