@jinyus 2023-11-20
Results on new VM:

Main: `18.78 ms | 238.77 ms  | 2.04 s `

This PR:
```
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 21.748ms
        Processing time (w/o IO): 21.589ms
        Processing time (w/o IO): 21.609ms
        Processing time (w/o IO): 23.335ms
        Processing time (w/o IO): 21.545ms
        Processing time (w/o IO): 21.564ms
        Processing time (w/o IO): 21.568ms
        Processing time (w/o IO): 22.458ms
        Processing time (w/o IO): 22.678ms
        Processing time (w/o IO): 21.602ms
        Processing time (w/o IO): 21.702ms
        Processing time (w/o IO): 21.476ms
        Processing time (w/o IO): 21.583ms
          Time (mean ± σ):      65.9 ms ±   0.7 ms    [User: 58.9 ms, System: 7.6 ms]
          Range (min … max):    65.3 ms …  67.3 ms    10 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 291.188ms
        Processing time (w/o IO): 293.344ms
        Processing time (w/o IO): 293.282ms
          Time (mean ± σ):     491.3 ms ±   1.1 ms    [User: 436.1 ms, System: 29.9 ms]
          Range (min … max):   490.5 ms … 492.1 ms    2 runs
         
D:

        Benchmark 1: ./related
        Processing time (w/o IO): 2517.82ms
        Processing time (w/o IO): 2517.11ms
        Processing time (w/o IO): 2526.59ms
          Time (mean ± σ):      3.175 s ±  0.036 s    [User: 3.000 s, System: 0.082 s]
          Range (min … max):    3.150 s …  3.201 s    2 runs
```